added API client, endpoints, and model schemas; refactored imports in components to use new structure
