+++ api-helpers/inputTypeSchemas/FriendshipScalarFieldEnumSchema.ts
