+++ api-helpers/inputTypeSchemas/GroupBalanceScalarFieldEnumSchema.ts
