+++ api-helpers/inputTypeSchemas/GroupScalarFieldEnumSchema.ts
