+++ api-helpers/inputTypeSchemas/GroupUserScalarFieldEnumSchema.ts
