+++ api-helpers/inputTypeSchemas/QueryModeSchema.ts
