+++ api-helpers/inputTypeSchemas/SplitTypeSchema.ts
