+++ components/providers.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { queryClient } from "@/api/client";
+import { queryClient } from "@/api-helpers/client";
 import { QueryClientProvider } from "@tanstack/react-query";
 import { AuthProvider } from "./AuthProvider";
 export function Providers({ children }: { children: React.ReactNode }) {
